+++ src/main/java/Employee.java
@@ -1,6 +1,6 @@
 public class Employee extends Person {
 
-	public Employee(int ID, int birthYear, String name, String surname, Contact contact) {
+	public Employee(long ID, int birthYear, String name, String surname, Contact contact) {
 		super(ID, birthYear, name, surname, contact);
 	}
